Commit message:
Subject:
"Added .gitkeep to preserve empty src directory"

Body:
"This commit introduces a new .gitkeep file in the src directory. Git doesn't track empty directories, and our project requires this src directory to remain in the repository, even if it's empty at the moment. By introducing a .gitkeep file, we can ensure that the src directory is preserved and committed. This will be beneficial once we begin adding source code files to this directory."